Validating the Discord Signature
The validation function we need to use
is defined on the `ed25519::Verifier`
trait. [`verify`](https://docs.rs/ed25519-dalek/1.0.1/ed25519_dalek/trait.Verifier.html#tymethod.verify)
will allow us to take a public key and
validate a message against a signature.

```shell
cargo add anyhow aws_lambda_events http
```

We'll write a function that accepts the headers
from the discord request, the body of the request
and the public key.

We'll also return an `anyhow::Result` as this
will make it easier to handle errors of different
types as well as construct errors.

```rust
pub fn validate_discord_signature(
    headers: &HeaderMap,
    body: &Body,
    pub_key: &PublicKey,
) -> anyhow::Result<()> {
    //
}
```

We'll take the validation in four steps:

1. Get the signature from the headers
1. Get the timestamp from the headers
1. Concatenate the timestamp with the body of the request
1. Validate the signature

Since we passed in a reference to the `HeaderMap`,
we can use `.get()` to get the value of a header if
it is set, or `None` otherwise. To communicate this
back to the calling function if it doesn't exist,
we'll use `.ok_or` which lets us specify the `Error`
for the `Err` and `?` which will return the error
if it exists.

To construct the error, we'll use the `anyhow!`
macro, which will build us an `anyhow::Error`
that fits the `anyhow::Result` in our return type.

After successfully getting the header, we need to
decode it from its hex representation to raw bytes.
If the decode fails, we'll add some context to
propogate back with the error.

Finally, because the `Signature::new` function
takes [a bytearray of length 64](https://docs.rs/ed25519/1.0.3/ed25519/struct.Signature.html#method.new),
we need to allocate an array of bytes (`u8`)
that is 64 bytes long. Then for each byte in
the decoded header, we'll set the corresponding
byte in the byte array.

After filling the byte array, we can create a
a new `Signature`. For additional exercises,
you can add additional checks to make sure the
header is of appropriate length before trying
to insert into the byte array.

```rust
let sig_ed25519 = {
    let header_signature = headers
        .get("X-Signature-Ed25519")
        .ok_or(anyhow!(
            "missing X-Signature-Ed25519 header"
        ))?;
    let decoded_header = hex::decode(header_signature)
        .context(
            "Failed to decode ed25519 signature header",
        )?;

    let mut sig_arr: [u8; 64] = [0; 64];
    for (i, byte) in
        decoded_header.into_iter().enumerate()
    {
        sig_arr[i] = byte;
    }
    Signature::new(sig_arr)
};
```

We can do the same thing that we did to get
the signature header to get the timestamp
header value.

```rust
let sig_timestamp =
    headers.get("X-Signature-Timestamp").ok_or(
        anyhow!("missing X-Signature-Timestamp header"),
    )?;
```

[`verify`](https://docs.rs/signature/1.3.0/signature/trait.Verifier.html#tymethod.verify)
takes a byte slice (not a byte array) and a signature,
so we can take advantage of iterators to chain
two series of bytes and collect them into a
`Vec<u8>`, which can then be passed as a slice
to the `verify` function.

```rust
if let Body::Text(body) = body {
    let content = sig_timestamp
        .as_bytes()
        .iter()
        .chain(body.as_bytes().iter())
        .cloned()
        .collect::<Vec<u8>>();

} else {
    Err(anyhow!("Invalid body type"))
}
```

Finally, we can use all of the pieces to verify
the signature. We'll add some more context to
any `Err` value that comes out of the verify
function and otherwise return `Ok(())`.

```rust
pub_key
    .verify(&content, &sig_ed25519)
    .context("Failed to verify the http body against the signature with the given public key")
```