Set up distributed tracing
When we put a service into production we are going
to have to debug it eventually. Doing this often
requires access to the logs we output from said
service. In our case we are outputting using the
`dbg!` macro to the console but we're losing
access to any errors in our validation function.

Why are we losing access? Partially because of
how lambdas work (and Netlify Functions is built
on top of AWS Lambda). Lambdas execute in two
stages. AWS will boot up your lambda code and
then hold it ready for events to come in. When
an event comes in, it will run your handler which
returns a response. In our case we're receiving
an http request and returning a response to the
user.

The second stage is the one we've been working
in so far: responding to the user. That means
that our errors are only going back to the user
as http responses and we don't have any insight
into what those errors are in our infrastructure.

The `lamedh_http::lambda` macro and the
`tokio::main` macro are both taking the main
function we've written and rewriting it.
`tokio::main` handles the async runtime
machinery, so it doesn't affect us in this case.
`lamedh_http::lambda` is making it so that we
only have to care about the second stage and is
hiding the construction of the first stage from
us.

While looking at this, we'll also want to upgrade
from arbitrary strings like what `dbg!` gives us,
to structured output that we can process later.

`tracing` is a framework for instrumenting Rust
programs to collect structured, event-based
diagnostic information.

To set up tracing, we'll remove the `lamedh`
macro and bootstrap the tracing infrastructure
in stage one, before all of our handlers run.

```rust
async fn main() -> Result<(), Error> {
    let my_subscriber = tracing_subscriber::fmt().finish();
    tracing::subscriber::set_global_default(my_subscriber)
        .expect("setting tracing default failed");

    run(lamedh_http::handler(handler)).await?;
    Ok(())
}
```

Then we'll add the `#[instrument]` macro to
the handler function. This will automatically
create a span when we enter and exit this function,
allowing us to automatically attach data to that
span. In this case we're adding the entire
`Request` as an example.

Inside of the `handler` function, we'll change our
`dbg!` macro to an `info!` event.

```rust
async fn handler(
    event: Request,
    _: Context,
) -> Result<impl IntoResponse, Error> {
    info!("in main");
    validate_discord_signature(
        event.headers(),
        event.body(),
        &PUB_KEY,
    )?;
    Ok("boop")
}
```

If we ship this to netlify and curl the function,
we'll see single-line formatted structured logs.
This will all headers and the request body
(because we told it to) as well as the message
from the `info!` event.

This formatting is a human-readable single-line
format. We can choose to make this even more
verbose, or even more compact and machine-readable
by outputting compact JSON instead. Since netlify
doesn't let us programmatically access our logs,
we'll leave the single-line human readable default.

```
8:01:29 PM: Apr 10 03:01:29.909  INFO handler{event=Request { method: GET, uri: https://pretest-discord-rust-slash-commands.netlify.app/.netlify/functions/hello-world, version: HTTP/1.1, headers: {"accept": "*/*", "client-ip": "67.180.63.210", "connection": "keep-alive", "user-agent": "curl/7.64.1", "via": "https/2 Netlify[9ad62796-12cd-413a-bdaf-bc1b36a37d47] (Netlify Edge Server)", "x-bb-ab": "0.054513", "x-bb-client-request-uuid": "9ad62796-12cd-413a-bdaf-bc1b36a37d47-5574061", "x-bb-ip": "67.180.63.210", "x-bb-loop": "1", "x-cdn-domain": "www.bitballoon.com", "x-country": "US", "x-datadog-parent-id": "5648543246425886663", "x-datadog-sampling-priority": "0", "x-datadog-trace-id": "6119018710333147135", "x-forwarded-for": "67.180.63.210", "x-forwarded-proto": "https", "x-nf-client-connection-ip": "67.180.63.210", "x-nf-request-id": "9ad62796-12cd-413a-bdaf-bc1b36a37d47-5574061", "host": "pretest-discord-rust-slash-commands.netlify.app"}, body: Empty }}: _in_progress_netlify_discord_slash_commands: in main
```

To add more instrumentation, we'll add another
`instrument` macro to our validation function.
In this case we'll skip the recording of the
function arguments. Additionally, we'll choose
to record any `Error` values that are returned
by the `Result` type our validation function
returns.

```rust
```

Now we can ship this to Netlify and view the new
logs, which will include an error. We can continue
to filter that log output based on log level (`INFO`
or `ERROR`) as well as more interesting data such
as client-ip.

---

In the future, we could add an additional subscriber
to the event data and ship our spans and events
somewhere that will let us query them, such as
Honeycomb. That is where distributed tracing comes
into play as we would be able to see not only the
spans, but all of the events laid out for us in
a visual tree layout.