Parsing JSON from a request body into Rust structs
The request that Discord sends our handler includes
JSON data. We need to parse this data into something
we can work with. The list of Discord's data models
can be found at https://discord.com/developers/docs/interactions/slash-commands#data-models-and-types

To serialize and deserialize data, serde is the defacto
standard. We need to install `serde` along with two
helper libraries: one for json and one for "C-like enums".

```shell
cargo add serde_json serde serde_repr
```

Together these will give us the minimal implementation
need to parse information out of the JSON string in the
request body from Discord.

We'll start with the absolute least effort solution:
letting serde do all the work. Using `from_slice` we
can let serde deserialize the request body into a
`serde_json::Value`, which is basically JSON as Rust
structs.

```rust
async fn handler(
    event: Request,
    _: Context,
) -> Result<impl IntoResponse, Error> {
    info!("in handler");
    validate_discord_signature(
        event.headers(),
        event.body(),
        &PUB_KEY,
    )?;
    let parsed: serde_json::Value =
        serde_json::from_slice(event.body())?;
    println!("{}", parsed);
    Ok("boop")
}
```

Printing this out with the display formatter will
give us the JSON our handler is receiving, which we
can then take into prettier to format if we want.

```json
{
  "application_id": "810835719329021973",
  "id": "830614249730605087",
  "token": "aW...q",
  "type": 1,
  "user": {
    "avatar": "21f8c900a6f25bd0da15d8f4ab66393b",
    "discriminator": "7892",
    "id": "103513724052082688",
    "public_flags": 0,
    "username": "chrisbiscardi"
  },
  "version": 1
}
```

Looking at the [Interaction](https://discord.com/developers/docs/interactions/slash-commands#interaction)
data model, we can see how the example JSON we've
received fits in and can start modeling it in
Rust more accurately.

The most important field for us is the `type`.
We'll build an `Interaction` struct to match the
naming Discord's docs use and derive `Deserialize`
for so that we can turn JSON into our struct and
`Debug` for convenience, so we can use `dbg!`.

`type` is a reserved word in Rust and while we
could use it anyway, we'll avoid that and use
`serde` to do a field-level rename when
deserializing so that `type` in JSON becomes
`interaction_type` in the `Interaction` struct.

Our second problem is that the `type` field
comes in as a number and each number represents
a different [`InteractionType`](https://discord.com/developers/docs/interactions/slash-commands#interaction-interactiontype).

```rust
struct Interaction {
    #[serde(rename = "type")]
    interaction_type: InteractionType,
}
```

This is very similar to an `enum` in Rust and with
the added help from `serde_repr` we can make this
true when deserializing. Similarly to before, we
derive `Deserialize_repr`, which is `serde_repr`'s
version of `Deserialize` and `Debug` for convenience.

We also use `repr` which is a native Rust construct
that isn't from any of the crates we added. `repr`
changes the data layout of our enum to store the
data in the same way a `u8` is stored.

```
    Deserialize_repr, Debug,
)]
pub enum InteractionType {
    Ping = 1,
    ApplicationCommand = 2,
}
```

With the structs laid out, we need to change the
type of the value returned from `serde_json::from_slice`.
After changing the type to `Interaction` Rust
handles everything else. We only need to say
that our data is structured in a specific way
and that it can be deserialized and we get
Rust versions of the JSON data.

```rust
let interaction: Interaction = serde_json::from_slice(event.body())?;
dbg!(interaction);
```

results in

```rust
[src/main.rs:51] interaction = Interaction {
    event_type: Ping,
}
```