Responding to a Ping Interaction
Now that we know we're getting a `Ping`, we'll
have to respond to that ping to complete the
interaction and save our interaction URL.

Let's start by constructing the types in the
[`InteractionResponse`](https://discord.com/developers/docs/interactions/slash-commands#interaction-response) section
of the docs. This is very similar to what we
did for the request types with the main
difference being that we're deriving `Serialize`
instead of `Deserialize`.

```rust
struct InteractionResponse {
    #[serde(rename = "type")]
    interaction_response_type: InteractionResponseType,
    data: Option<InteractionApplicationCommandCallbackData>,
}

pub enum InteractionResponseType {
    Pong = 1,
    // Acknowledge = 2,
    // ChannelMessage = 3,
    ChannelMessageWithSource = 4,
    ACKWithSource = 5,
}

struct InteractionApplicationCommandCallbackData {
    tts: Option<bool>,
    content: Option<String>,
    // embeds
    // allowed_mentions
    flags: Option<usize>,
}
```

We have to respond differently to Discord's `Ping`s
compared to `ApplicationCommand`s, so we can match
on the interaction type. If we get a `Ping`, we
will return an `InteractionResponse` with a type of
`Pong`.

```rust
match interaction.interaction_type {
    InteractionType::Ping => Ok(InteractionResponse {
        interaction_response_type:
            InteractionResponseType::Pong,
        data: None,
    }),
    InteractionType::ApplicationCommand => {
        todo!("implement ApplicationCommand")
    }
}
```

This leaves us with the need to implement `IntoResponse`
for our `InteractionResponse` type.

```
error[E0277]: the trait bound `InteractionResponse: IntoResponse` is not satisfied
  --> src/main.rs:45:13
   |
45 | ) -> Result<impl IntoResponse, Error> {
   |             ^^^^^^^^^^^^^^^^^ the trait `IntoResponse` is not implemented for `InteractionResponse`
```

Which we can do by taking advantage of serde_json
and http's `Response::builder`. We use `expect`
here because it should never fail.

```rust
impl IntoResponse for InteractionResponse {
    fn into_response(self) -> Response<Body> {
        Response::builder()
            .header(CONTENT_TYPE, "application/json")
            .body(
                serde_json::to_string(&self)
                    .expect("unable to serialize serde_json::Value")
                    .into(),
            )
            .expect("unable to build http::Response")
    }
}
```

Before we ship, we also should take note of
the additional requirement that failing
signatures should be responded to with 401s.

We can change our validation call to use an
if-let. We're already recording the error
using tracing, so we can ignore that argument
completely.

We'll use `http::Response::builder` to build
up a 401 response with some arbitrary text.

```rust
if let Err(_) = validate_discord_signature(
    event.headers(),
    event.body(),
    &PUB_KEY,
) {
    return Ok(Response::builder()
        .status(StatusCode::UNAUTHORIZED)
        .body(Body::Text(
            "invalid request signature".to_string(),
        ))
        .expect("response builder failed"));
};
```

This leads us to the following error.

```
error[E0308]: mismatched types
  --> src/main.rs:63:37
   |
63 |           InteractionType::Ping => Ok(InteractionResponse {
   |  _____________________________________^
64 | |             interaction_response_type:
65 | |                 InteractionResponseType::Pong,
66 | |             data: None,
67 | |         }),
   | |_________^ expected struct `Response`, found struct `InteractionResponse`
   |
   = note: expected struct `Response<Body>`
              found struct `InteractionResponse`
```

We get this type mismatch because in different
branches we try to return different types. The
way `impl IntoResponse` works is that we don't
have to specify the concrete type we're using
up front, which means it can be
`InteractionResponse` or `Response<Body>`, but
only one of them can fill the `impl IntoResponse`
slot for the return value of the function.

We have a couple of options to solve this. The
easiest one is to call `into_response` ourselves.
We've defined it on the `InteractionResponse`
type, so it'll be there for us on every value
of that type.

```rust
match interaction.interaction_type {
    InteractionType::Ping => Ok(InteractionResponse {
        interaction_response_type:
            InteractionResponseType::Pong,
        data: None,
    }
    .into_response()),
    InteractionType::ApplicationCommand => {
        todo!("implement ApplicationCommand")
    }
}
```

Another way to solve this problem is to
create an enum whose variants are all of
the types we want to allow responding with.
We can then implement `IntoResponse` for
that enum and use it to wrap all of our
return values. Because the enum is one
type it will work although it also requires
more ceremony so is more applicable to
larger projects.