Create a Discord slash command with curl
There are two places to create a Discord slash
command. One is global, which applies to all
guilds/servers and the other is local, which
applies to specific guilds/servers. The key
for us is that creating commands in specific
servers immediately updates and allows us to
test those commands while creating global
commands make us deal with some caching for
those commands.

We create the following file to define our
"self-assign a role" command.

```json
{
  "name": "role",
  "description": "Request a role, some roles are self-assignable.",
  "options": [
    {
      "name": "i-want-to-be-a",
      "description": "Choose the role you want.",
      "type": 8,
      "required": true
    }
  ]
}
```

and we can use curl to create the command

```shell
curl -XPOST \
        -H "Content-Type: application/json" \
        -H "Authorization: Bot $DISCORD_BOT_TOKEN" \
        -d @./interactions/role.json \
        https://discord.com/api/v8/applications/<app-id>/guilds/<guild-id>/commands
```