Using Methods and impls to simplify boilerplate
Constructing a full `InteractionResponse` struct
for each response gets fairly tiresome when we
are using the same options for every single
reply. Using `impl` blocks we can define methods
on the `InteractionResponse` object that allow
us to supply a string for a reply and abstracting
the rest of the boilerplate away.

```rust
impl InteractionResponse {
    fn reply(content: String) -> InteractionResponse {
        InteractionResponse{
            interaction_response_type: InteractionResponseType::ChannelMessageWithSource,
            data: Some(InteractionApplicationCommandCallbackData {
                tts: None,
                content: Some(content),
                flags: None
            })
        }
    }
}
```

Using this method involves calling it from the
`InteractionResponse` type. This allows us to
clean up a bunch of the code we're not changing
betwen our calls, making the code stand out and
not be hidden inside of that boilerplate.

```rust
Ok(InteractionResponse::reply(
    "received role command".to_string(),
)
.into_response())
```